# artifacts/api-server/openapi.yaml
openapi: 3.0.3
info:
  title: ISM Finance API
  version: 1.2.0
  description: API REST do SaaS financeiro com painel JWT e integração Power BI (v1 token).
servers:
  - url: http://localhost:3000/api
    description: Local
...
paths:
...
  /relatorios/fechamento-mensal:
    get:
      tags: [Relatorios]
      operationId: getRelatoriosFechamentoMensal
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: mes
          required: true
          schema: { type: integer }
        - in: query
          name: ano
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Fechamento mensal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /relatorios/dre:
    get:
      tags: [Relatorios]
      operationId: getRelatoriosDre
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ano
          schema: { type: integer }
        - in: query
          name: regime
          schema: { type: string, enum: [competencia, caixa], default: competencia }
      responses:
        "200":
          description: Relatório DRE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /relatorios/fluxo-caixa:
    get:
      tags: [Relatorios]
      operationId: getRelatoriosFluxoCaixa
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Fluxo de caixa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /relatorios/metas:
    get:
      tags: [Relatorios]
      operationId: getRelatoriosMetas
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Relatório de metas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /relatorios/contabil-fiscal:
    get:
      tags: [Relatorios]
      operationId: getRelatoriosContabilFiscal
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: data_inicio
          schema: { type: string, format: date }
        - in: query
          name: data_fim
          schema: { type: string, format: date }
        - in: query
          name: conta_id
          schema: { type: integer }
        - in: query
          name: tipo
          schema: { type: string, enum: [CP, CR, ambos], default: ambos }
      responses:
        "200":
          description: Relatório contábil/fiscal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerV1Token:
      type: http
      scheme: bearer
      description: Token estático validado na tabela `tokens_api`.
...
  schemas:
    ApiErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    ApiEnvelopeGeneric:
      type: object
      properties:
        data: {}
        meta:
          nullable: true
          type: object
          additionalProperties: true
        errors:
          nullable: true
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorItem"
...